Make SocketAddress convertible from/to IP address and port pair in string form
Internet addresses come from all sorts of places and in various forms, but the most common one is the `\(ipAddress):\(port ?? 0)`, take `127.0.0.1:80` as an example.

Being able to switch back and forth between the `SocketAddress` and the aforementioned string form seems to be small enough to fit in `NIO`, this way also avoiding sprouts of monkey-patches across the dependent apps and packages.